Reactor Command Gateway Tests
